# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Magic Garden API
  version: 2.0.0
  description: |
    **Unofficial API** for Magic Garden game data that dynamically extracts and provides game content.

    ## Overview

    This API provides two types of data:
    - **Static data**: Extracted from the game's minified bundle (plants, pets, items, etc.)
    - **Live data**: Real-time data from the game's WebSocket (shops, weather)

    ## Quick Start

    ### Get All Game Data
    ```bash
    curl https://mg-api.ariedam.fr/data
    ```

    ### Get Specific Plant
    ```bash
    curl https://mg-api.ariedam.fr/data/plants | jq '.Carrot'
    ```

    ### Stream Live Updates
    ```javascript
    const eventSource = new EventSource('https://mg-api.ariedam.fr/live/stream');
    eventSource.addEventListener('weather', (event) => {
      const data = JSON.parse(event.data);
      console.log('Weather:', data.weather);
    });
    eventSource.addEventListener('shops', (event) => {
      const shops = JSON.parse(event.data);
      console.log('Seed shop:', shops.seed);
    });
    ```

    ## Key Features

    - **Auto-updating**: Data automatically refreshes from the game bundle
    - **Real-time**: WebSocket connection for live shop and weather data
    - **Sprites included**: All data includes sprite URLs ready to use
    - **Server-Sent Events**: Real-time streaming via SSE
    - **Smart caching**: Optimized performance with automatic cache invalidation
...

    ## Data Structure

...

    ## Rate Limiting
...
servers:
...
paths:
...
  /assets/animations:
    get:
      tags:
        - Assets
      summary: Catalog of animated pet loops (WebP/GIF)
      description: |
        Pets are the only creatures the game renders as vectors rather than
        sprites, so beyond the still PNG served by `/assets/sprites/pets/*`
        this API can serve them **animated**: one looping WebP (or GIF) per
        species and per state.

        Loops are pre-rendered whenever the game ships a new pet file - never
        on request - and served as plain files. Drop the URL in an `<img>` tag
        and it plays; there is no runtime to embed.

        **Clips:**
        - `idle` - the resting loop, 7 s
        - `walk` - one step cycle, ~0.6 s
        - `eat` - eating, 1 s
        - `sleep` - sleeping, 8 s

        Not every species exposes every clip, and the two weather variants
        (`FireHorseActive`, `ThunderWolfActive`) only ship `idle`.

        **Filters:** `cat`, `name` (exact species), `clip`, `search`
        (case-insensitive substring on the species name).

        **Response notes:**
        - Every loop is rendered to the same subject height, so species stay
          visually comparable - a snail and a horse come out the same size.
        - `url` is the link, in the primary format (WebP by default); `format`
          names it, and any additional format appears under its own key
          (`gif`).
        - `anchor` places the sprite like the game does: `x` is always 0.5,
          `y` marks the ground line under the pet's feet.
        - `source.generatedAt` tells you when the loops were last rendered, and
          `source.riveUrl` which game file they came from.
        - Not to be confused with the `animations` **sprite** category
          (`/assets/sprites/animations/*.png`), which holds still frames from
          the game's atlases.
      parameters:
        - name: cat
          in: query
          required: false
          schema:
            type: string
            enum: [pets]
          description: Only return animations from this category
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Exact species name (case-insensitive)
          example: "Chicken"
        - name: clip
          in: query
          required: false
          schema:
            type: string
            enum: [idle, walk, eat, sleep]
          description: Only return this clip
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive substring match on species names
          example: "horse"
      responses:
        '200':
          description: Animation catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 118
                  baseUrl:
                    type: string
                    example: "https://mg-api.ariedam.fr"
                  categories:
                    type: array
                    items:
                      type: string
                    example: ["pets"]
                  formats:
                    type: array
                    items:
                      type: string
                    description: Formats currently generated
                    example: ["webp"]
                  source:
                    type: object
                    properties:
                      riveUrl:
                        type: string
                        description: Game file the loops were rendered from
                      generatedAt:
                        type: string
                        format: date-time
                  animations:
                    type: object
                    description: Animations grouped by category
                    additionalProperties:
                      type: array
                      items:
                        $ref: '#/components/schemas/PetAnimation'
              examples:
                filtered:
                  summary: One species (?name=Chicken)
                  value:
                    count: 4
                    baseUrl: "https://mg-api.ariedam.fr"
                    categories: ["pets"]
                    formats: ["webp"]
                    source:
                      riveUrl: "https://magicgarden.gg/runtime-assets/pets.bfec8b856f6e2b0e4a83.riv"
                      generatedAt: "2026-08-04T14:31:07.412Z"
                    animations:
                      pets:
                        - category: "pets"
                          name: "Chicken"
                          clip: "idle"
                          timeline: "Pet_Idle"
                          url: "https://mg-api.ariedam.fr/assets/animations/pets/Chicken_idle.webp?v=824"
                          format: "webp"
                          width: 212
                          height: 257
                          frames: 210
                          fps: 30
                          durationMs: 7000
                          anchor:
                            x: 0.5
                            y: 1
        '400':
          description: Unknown category in `cat`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /assets/animations/{category}/{file}:
    get:
      tags:
        - Assets
      summary: Download an animated loop
      description: |
        Download one looping animation as an animated WebP or GIF. The filename
        is `<Species>_<clip>.<webp|gif>`.

        WebP is the format to prefer: 8-bit alpha (these sprites have soft
        edges and translucent shadows) and about half the bytes of the
        equivalent GIF. GIF is only generated when the server is configured for
        it, and pays for compatibility with a 256-colour palette and 1-bit
        transparency.

        **Examples:**
        - `/assets/animations/pets/Chicken_idle.webp` - chicken, idle loop
        - `/assets/animations/pets/Horse_walk.webp` - horse, one step cycle
        - `/assets/animations/pets/ThunderWolfActive_idle.webp` - thunder wolf under active weather

        **HTML:**
        ```html
        <img src="https://mg-api.ariedam.fr/assets/animations/pets/Chicken_idle.webp" alt="Chicken">
        ```

        The file loops forever on its own - no JavaScript, no Rive runtime.
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
            enum: [pets]
          example: "pets"
        - name: file
          in: path
          required: true
          description: "`<Species>_<clip>.<webp|gif>`"
          schema:
            type: string
            pattern: '^[A-Za-z0-9-]+_[a-z0-9]+\.(webp|gif)$'
          example: "Chicken_idle.webp"
      responses:
        '200':
          description: Animated image
          content:
            image/webp:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid category or filename
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Animation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                notFound:
                  summary: Species has no such clip
                  value:
                    error:
                      code: "NOT_FOUND"
                      message: "Animation not found: pets/Chicken_dance.webp"
...
components:
  schemas:
...
    PetAnimation:
      type: object
      description: Catalog entry for one looping animation
      properties:
        category:
          type: string
          example: "pets"
        name:
          type: string
          description: Artboard name (species, or weather variant)
          example: "Chicken"
        clip:
          type: string
          enum: [idle, walk, eat, sleep]
          example: "idle"
        timeline:
          type: string
          description: Rive timeline the loop was rendered from
          example: "Pet_Idle"
        url:
          type: string
          format: uri
          example: "https://mg-api.ariedam.fr/assets/animations/pets/Chicken_idle.webp?v=824"
        format:
          type: string
          description: Format `url` points to
          example: "webp"
        gif:
          type: string
          format: uri
          description: Only present when the server also generates GIFs
        width:
          type: integer
          example: 212
        height:
          type: integer
          example: 257
        frames:
          type: integer
          example: 210
        fps:
          type: number
          example: 30
        durationMs:
          type: integer
          example: 7000
        anchor:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
        bytes:
          type: object
          description: File size per format
          additionalProperties:
            type: object
            properties:
              bytes:
                type: integer
...
    Error:
      type: object
      description: Standard error response format
      properties:
        error:
          type: object
          description: Error details
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "SPRITE_NOT_FOUND"
            message:
              type: string
              description: Human-readable error message
              example: "The requested sprite was not found"
            details:
              type: object
              description: Additional error context (varies by error type)
              additionalProperties: true
              example:
                category: "seeds"
                filename: "NonExistent.png"
      example:
        error:
          code: "SPRITE_NOT_FOUND"
          message: "The requested sprite was not found"
          details:
            category: "seeds"
            filename: "NonExistent.png"
...